replace int with int64_t

+++ src/Open3D/Container/Broadcast.cpp
@@ -62,10 +62,10 @@ SizeVector BroadcastedShape(const SizeVector& left_shape,
                           left_shape, right_shape);
     }
 
-    int left_ndim = left_shape.size();
-    int right_ndim = right_shape.size();
-    int shorter_ndim = std::min(left_ndim, right_ndim);
-    int longer_ndim = std::max(left_ndim, right_ndim);
+    int64_t left_ndim = left_shape.size();
+    int64_t right_ndim = right_shape.size();
+    int64_t shorter_ndim = std::min(left_ndim, right_ndim);
+    int64_t longer_ndim = std::max(left_ndim, right_ndim);
 
     if (left_ndim == 0) {
         return right_shape;
@@ -76,11 +76,11 @@ SizeVector BroadcastedShape(const SizeVector& left_shape,
 
     SizeVector broadcasted_shape(longer_ndim, 0);
     // Checked from right to left
-    for (int ind = 0; ind < longer_ndim; ind++) {
-        int left_ind = left_ndim - longer_ndim + ind;
-        int right_ind = right_ndim - longer_ndim + ind;
-        int left_dim = left_ind >= 0 ? left_shape[left_ind] : 0;
-        int right_dim = right_ind >= 0 ? right_shape[right_ind] : 0;
+    for (int64_t ind = 0; ind < longer_ndim; ind++) {
+        int64_t left_ind = left_ndim - longer_ndim + ind;
+        int64_t right_ind = right_ndim - longer_ndim + ind;
+        int64_t left_dim = left_ind >= 0 ? left_shape[left_ind] : 0;
+        int64_t right_dim = right_ind >= 0 ? right_shape[right_ind] : 0;
         broadcasted_shape[ind] = std::max(left_dim, right_dim);
     }
     return broadcasted_shape;
